fix: Avoid loading datasets just for reindexing
The reindexing process was accidentally deserialising the large feather
files from disk as it went, causing way more memory to be used than
necessary.

This change iterates over the metadata files only when reindexing,
speeding it up quite a lot.

As part of the change, it brings the `primary_key` attribute into the
`TableMeta` class. This means we also need to set it and update it
on `set_index()`, so some minor changes done there too.

https://github.com/owid/etl/issues/34