fix #21: allow loading from local catalogs
The `LocalCatalog()` would firstly not correctly work unless it had
been indexed. Secondly, it would fail on `load()` of a table because
the `_base_uri` had not been set.

This fix automatically reindexes a local catalog if no index is present,
and secondly fixes the load method so that you can properly load local
datasets.