fix: Avoid reloading the from disk on reindex
When you create a local catalog, it doesn't yet know how to support
find() operations unless there's an index for the catalog. When no
index existed, the catalog would rebuild one on the spot, save it
to disk and then reload it from disk.

This fix prevents only loads it from disk if it's already there. It also
ensures the `_base_uri` is set on `reindex()`.