fix: could not use a series with apply
We were getting an error with a statement like:

```
df['a'] = df.a.apply(...)
```

It turns out that this is because the new series lost its `_name`, and
without that it cannot be linked with its metadata properly.

This fix simply adds `_name` to the list of fields that Pandas knows to
copy each time you slice or change a series, meaning the name does not
get lost any more with these transforms.