feature: add parquet support and render to multiple files
Parquet is an efficient columnar format like Feather. We want to support
both and render to both since some downstream consumers only support one
of these formats, and Parquet is better supported than Feather
generally.

Parquet supports a schema with additional metadata; we take advantage of
that and serialise our table metadata into the schema on save, restoring
it on write.

By default, we also change Dataset so that when you add a table, it
serialises the table to all available formats including CSV unless
specifically restricted.